Rename ethereum watcher module to adapter
